feat: refactor profile preferences component into modular sections
